requirements.txt
python-dotenv
oracledb
pandas
psycopg2-binary
argon2-cffi